requirements.txt
python-telegram-bot[webhooks]==20.7
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0